refactor to metastructure package
