+++ bim2sim/export/modelica/standardlibrary.py
@@ -1,7 +1,7 @@
 ﻿"""Modul containing model representations from the Modelica Standard Library"""
 from bim2sim.export import modelica
 from bim2sim.kernel import aggregation
-from bim2sim.kernel.elements import hvac
+from bim2sim.metastructure import hvac
 from bim2sim.kernel.units import ureg
