--- bim2sim/kernel/elements/__init__.py
@@ -1,2 +1,0 @@
-
-schema = 'IFC4'
