+++ bim2sim/plugins/PluginEnergyPlus/bim2sim_energyplus/task/ep_geom_preprocessing.py
@@ -24,9 +24,8 @@ from OCC.Core.TopExp import TopExp_Explorer
 from OCC.Core.TopoDS import topods_Face, TopoDS_Shape
 from OCC.Core.gp import gp_Pnt, gp_Trsf, gp_XYZ, gp_Vec
 
-from bim2sim.decision import BoolDecision, DecisionBunch
-from bim2sim.kernel.elements.bps import ExternalSpatialElement, SpaceBoundary, \
-    ThermalZone, SpaceBoundary2B
+from bim2sim.metastructure.bps import ExternalSpatialElement, SpaceBoundary, \
+    SpaceBoundary2B
 from bim2sim.task.base import ITask
 from bim2sim.task.common.inner_loop_remover import convex_decomposition, \
     is_convex_no_holes, is_convex_slow
