+++ bim2sim/plugins/PluginHKESim/bim2sim_hkesim/__init__.py
@@ -3,8 +3,6 @@
 Holds logic to run a simulation based on prepared ifc data
 """
 from bim2sim.export.modelica import standardlibrary
-from bim2sim.kernel.element import Material
-from bim2sim.kernel.elements import hvac as hvac_elements
 from bim2sim.plugins import Plugin
 from bim2sim.task import base, common, hvac
 from bim2sim.simulation_settings import PlantSimSettings
