+++ bim2sim/plugins/PluginLCA/bim2sim_lca/__init__.py
@@ -2,8 +2,6 @@
 
 Holds logic to export LCA relevant information and quantities based on ifc data
 """
-from bim2sim.kernel.element import Material
-from bim2sim.kernel.elements import hvac as hvac_elements, bps as bps_elements
 from bim2sim.plugins import Plugin
 from bim2sim.plugins.PluginLCA.bim2sim_lca.task.export_lca import ExportLCA
 from bim2sim.task import common, bps
