+++ bim2sim/plugins/PluginTEASER/bim2sim_teaser/task/export_teaser.py
@@ -12,7 +12,7 @@ from teaser.logic.buildingobjects.buildingphysics.rooftop import Rooftop
 from teaser.logic.buildingobjects.buildingphysics.window import Window
 from teaser.project import Project
 
-from bim2sim.kernel.element import ProductBased
+from bim2sim.metastructure.__init__ import ProductBased
 from bim2sim.task.base import ITask
 from bim2sim.utilities.common_functions import filter_instances
