+++ bim2sim/task/bps/check_ifc.py
@@ -1,7 +1,2 @@
-from ifcopenshell import file
 
-from bim2sim.kernel.elements import bps
-from bim2sim.kernel.ifc2python import get_layers_ifc
-from bim2sim.task.base import Playground
-from bim2sim.task.common.common_tasks import CheckIfc
