+++ bim2sim/task/bps/enrich_material.py
@@ -2,12 +2,11 @@ import ast
 import re
 
 from bim2sim.decision import ListDecision, DecisionBunch
-from bim2sim.kernel.element import Material
-from bim2sim.kernel.elements.bps import Layer, LayerSet, Building
+from bim2sim.metastructure.__init__ import Material
+from bim2sim.metastructure.bps import Layer, LayerSet, Building
 from bim2sim.task.base import ITask
 from bim2sim.utilities.common_functions import get_material_templates, \
     translate_deep, filter_instances, get_type_building_elements
-from bim2sim.simulation_settings import BaseSimSettings
 
 
 class EnrichMaterial(ITask):
