+++ bim2sim/task/bps/verify.py
@@ -1,10 +1,9 @@
-from bim2sim.kernel.element import Material
-from bim2sim.kernel.elements.bps import BPSProductWithLayers, LayerSet, Layer
+from bim2sim.metastructure.__init__ import Material
+from bim2sim.metastructure.bps import BPSProductWithLayers, LayerSet, Layer
 from bim2sim.kernel.units import ureg
 from bim2sim.task.base import ITask
 from bim2sim.utilities.common_functions import all_subclasses, filter_instances
 from bim2sim.utilities.types import LOD
-from bim2sim.simulation_settings import BaseSimSettings
 
 
 class Verification(ITask):
