+++ bim2sim/task/hvac/check_ifc.py
@@ -1,7 +1,2 @@
-from ifcopenshell.entity_instance import entity_instance
 
-import bim2sim.task.common as cm
-from bim2sim.kernel.elements import hvac
-from bim2sim.kernel.ifc2python import get_ports
-from bim2sim.task.base import Playground
