+++ test/unit/kernel/aggregation/test_consumer_distributor.py
@@ -1,7 +1,7 @@
 import unittest
 
 from bim2sim.kernel import aggregation
-from bim2sim.kernel.elements import hvac
+from bim2sim.metastructure import hvac
 from bim2sim.kernel.hvac.hvac_graph import HvacGraph
 from test.unit.kernel.helper import SetupHelperHVAC
