+++ test/unit/kernel/test_attribute.py
@@ -3,13 +3,13 @@
 import unittest
 
 from bim2sim.decision import DecisionBunch, RealDecision
-from bim2sim.kernel import element
+from bim2sim.metastructure import ProductBased
 from bim2sim.kernel.attribute import Attribute
 from bim2sim.kernel.units import ureg
 from test.unit.kernel.helper import SetupHelperHVAC
 
 
-class TestElement(element.ProductBased):
+class TestElement(ProductBased):
     ifc_types = {}
 
     attr1 = Attribute(
